Update 02-local-development.md (#232)
* Update 02-local-development.md

When pull request #1517 was merged to lemmy-ui, which I have a link to [here](https://github.com/LemmyNet/lemmy-ui/commit/d0dff77377a84611ab5cf624123e84c95aedf1ff), the dev command to run the local frontend was changed from `yarn start` to `yarn build:dev --watch`. This updates the documents to reflect that.

* Update 02-local-development.md

This is a better shorthand for the command, it does the same thing but is shorter.